refactor: rename semantic::context module to semantic::decl
