+++ crates/prql-compiler/src/semantic/lowering.rs
@@ -13,13 +13,13 @@ use crate::ir::pl::{self, Ident, Lineage, LineageColumn, PlFold, QueryDef, Tuple
 use crate::ir::rq::{
     self, CId, RelationColumn, RelationLiteral, RelationalQuery, TId, TableDecl, Transform,
 };
-use crate::semantic::context::TableExpr;
+use crate::semantic::decl::TableExpr;
 use crate::semantic::module::Module;
 use crate::semantic::write_pl;
 use crate::utils::{toposort, IdGenerator};
 use crate::COMPILER_VERSION;
 
-use super::context::{self, DeclKind};
+use super::decl::{self, DeclKind};
 use super::{RootModule, NS_DEFAULT_DB};
 
 /// Convert AST into IR and make sure that:
@@ -162,8 +162,8 @@ impl Lowerer {
         }
     }
 
-    fn lower_table_decl(&mut self, table: context::TableDecl, fq_ident: Ident) -> Result<()> {
-        let context::TableDecl { ty, expr } = table;
+    fn lower_table_decl(&mut self, table: decl::TableDecl, fq_ident: Ident) -> Result<()> {
+        let decl::TableDecl { ty, expr } = table;
 
         // TODO: can this panic?
         let columns = ty.unwrap().into_relation().unwrap();
@@ -947,12 +947,12 @@ fn validate_take_range(range: &Range<rq::Expr>, span: Option<Span>) -> Result<()
 struct TableExtractor {
     path: Vec<String>,
 
-    tables: Vec<(Ident, context::TableDecl)>,
+    tables: Vec<(Ident, decl::TableDecl)>,
 }
 
 impl TableExtractor {
     /// Finds table declarations in a module, recursively.
-    fn extract(root_module: &Module) -> Vec<(Ident, context::TableDecl)> {
+    fn extract(root_module: &Module) -> Vec<(Ident, decl::TableDecl)> {
         let mut te = TableExtractor::default();
         te.extract_from_module(root_module);
         te.tables
@@ -982,9 +982,9 @@ impl TableExtractor {
 /// are not needed for the main pipeline. To do this, it needs to collect references
 /// between pipelines.
 fn toposort_tables(
-    tables: Vec<(Ident, context::TableDecl)>,
+    tables: Vec<(Ident, decl::TableDecl)>,
     main_table: &Ident,
-) -> Vec<(Ident, context::TableDecl)> {
+) -> Vec<(Ident, decl::TableDecl)> {
     let tables: HashMap<_, _, RandomState> = HashMap::from_iter(tables);
 
     let mut dependencies: Vec<(Ident, Vec<Ident>)> = Vec::new();
